Use work counters for grapheme scaling tests
