+++ safere-crosscheck/pom.xml
@@ -120,6 +120,7 @@
                         <exclude name="DfaTest.java"/>
                         <exclude name="EmptyOpTest.java"/>
                         <exclude name="EnginePathEquivalenceTest.java"/>
+                        <exclude name="GraphemeLinearTimeTest.java"/>
                         <exclude name="InstOpTest.java"/>
                         <exclude name="InstTest.java"/>
                         <exclude name="MatcherDeferredCaptureStateTest.java"/>
