+++ safere/src/main/java/org/safere/Dfa.java
@@ -1117,6 +1117,7 @@ final class Dfa {
 
     int pos = startPos;
     while (pos <= textLen) {
+      WorkCounter.record();
       int cp;
       int nextPos;
       int cls;
@@ -1271,6 +1272,7 @@ final class Dfa {
 
     int pos = endPos;
     while (pos >= startLimit) {
+      WorkCounter.record();
       int cp;
       int prevPos;
       int cls;
