+++ safere/src/main/java/org/safere/Matcher.java
@@ -390,6 +390,7 @@ public final class Matcher implements MatchResult {
     // Scan every code point.
     int i = 0;
     while (i < len) {
+      WorkCounter.record();
       int cp = text.codePointAt(i);
       if (cp < 64) {
         if ((b0 & (1L << cp)) == 0) {
@@ -421,6 +422,7 @@ public final class Matcher implements MatchResult {
     int i = fromIndex;
     int len = text.length();
     while (i < len) {
+      WorkCounter.record();
       int cp = text.codePointAt(i);
       if (charClassContains(ranges, b0, b1, cp)) {
         return applyEngineResult(new FullMatchResult(new int[] {i, i + Character.charCount(cp)}));
@@ -441,6 +443,7 @@ public final class Matcher implements MatchResult {
     int i = Math.max(0, fromIndex);
     int len = text.length();
     while (i < len) {
+      WorkCounter.record();
       int cp = text.codePointAt(i);
       if (charClassContains(ranges, b0, b1, cp)) {
         return true;
@@ -1214,6 +1217,7 @@ public final class Matcher implements MatchResult {
       if (parentPattern.prefixFoldCase()) {
         idx = indexOfIgnoreCase(text, literal, searchFrom);
       } else {
+        WorkCounter.record(Math.max(0, text.length() - searchFrom));
         idx = text.indexOf(literal, searchFrom);
       }
       if (idx < 0) {
@@ -1290,6 +1294,7 @@ public final class Matcher implements MatchResult {
       if (parentPattern.prefixFoldCase()) {
         idx = indexOfIgnoreCase(text, prefix, searchFrom);
       } else {
+        WorkCounter.record(Math.max(0, text.length() - searchFrom));
         idx = text.indexOf(prefix, searchFrom);
       }
       if (idx < 0) {
@@ -1646,6 +1651,7 @@ public final class Matcher implements MatchResult {
   private boolean findKeywordAlternation(
       Pattern.KeywordAlternation keywordAlternation, int startPos, int ncap) {
     for (int i = Math.max(0, startPos); i < text.length(); i++) {
+      WorkCounter.record();
       char ch = text.charAt(i);
       if (ch < 128
           && keywordAlternation.firstAscii[asciiLower(ch)]
@@ -1695,6 +1701,7 @@ public final class Matcher implements MatchResult {
     int prefixLen = prefix.length();
     int limit = text.length() - prefixLen;
     for (int i = fromIndex; i <= limit; i++) {
+      WorkCounter.record();
       if (regionMatchesAsciiIgnoreCase(text, i, prefix, 0, prefixLen)) {
         return i;
       }
@@ -1712,6 +1719,7 @@ public final class Matcher implements MatchResult {
       return false;
     }
     for (int i = 0; i < length; i++) {
+      WorkCounter.record();
       if (asciiLower(text.charAt(textOffset + i)) != asciiLower(prefix.charAt(prefixOffset + i))) {
         return false;
       }
@@ -1726,6 +1734,7 @@ public final class Matcher implements MatchResult {
    */
   private static int indexOfCharClass(String text, boolean[] asciiMap, int fromIndex) {
     for (int i = fromIndex; i < text.length(); i++) {
+      WorkCounter.record();
       char ch = text.charAt(i);
       if (ch < 128 && asciiMap[ch]) {
         return i;
@@ -1738,6 +1747,7 @@ public final class Matcher implements MatchResult {
       String text, Pattern.StartAcceleration acceleration, int fromIndex, boolean unixLines) {
     int start = Math.max(0, fromIndex);
     for (int i = start; i < text.length(); i++) {
+      WorkCounter.record();
       if (matchesStartAcceleration(text, i, acceleration, unixLines)) {
         return i;
       }
