+++ safere/src/main/java/org/safere/Nfa.java
@@ -560,6 +560,7 @@ final class Nfa {
 
     int engineEndPos = context.engineEndPos();
     for (int pos = start; pos < engineEndPos + 2; pos++) {
+      WorkCounter.record();
       if (prog.hasGraphemeSemantics()) {
         mergeDelayedQueue(delayedGrapheme, pos, runq);
       } else {
@@ -864,6 +865,7 @@ final class Nfa {
       int matchPos,
       int nextPos) {
     for (int threadIndex = 0; threadIndex < rq.size; threadIndex++) {
+      WorkCounter.record();
       NfaThread t = rq.threads[threadIndex];
       int id = t.id;
       int[] capture = t.capture;
